Cast time dim values to int when calculating corresponding time index (#656)

+++ data/netcdf_data.py
@@ -108,7 +108,7 @@ class NetCDFData(Data):
             [int or list] -- Time index(es).
         """
 
-        time_var = self.time_variable
+        time_var = self.time_variable.astype(np.int)
 
         result = np.nonzero(np.isin(time_var, timestamp))[0]
